chore(error): extract manage-error from controllers
